Make an inconsistent "fix" for ReLU when the input is a matrix
The ReLU and Softmax operations currently return a jacobian given vector
inputs + outputs. This doesn't work when the ReLU input is a matrix as
is the case for the Conv2D layer.

This commit makes an inconsistent "fix" where ReLU returns a matrix if
the input is a matrix. This matches the output of eg. Tensorflow's
gradient for `relu(matrix_tensor)`. Fixing this to make the output of
activation ops' `gradient` functions consistent is TBD.